+++ src/terminal/VTType.h
@@ -14,6 +14,7 @@
 #pragma once
 
 #include <string>
+#include <fmt/format.h>
 
 namespace terminal {
 
@@ -74,3 +75,30 @@ std::string to_string(DeviceAttributes v);
 std::string to_params(DeviceAttributes v);
 
 }  // namespace terminal
+
+namespace fmt // {{{
+{
+    template <>
+    struct formatter<terminal::VTType> {
+        template <typename ParseContext>
+        constexpr auto parse(ParseContext& ctx) { return ctx.begin(); }
+        template <typename FormatContext>
+        auto format(const terminal::VTType _id, FormatContext& ctx)
+        {
+            switch (_id)
+            {
+                case terminal::VTType::VT100: return format_to(ctx.out(), "VT100");
+                case terminal::VTType::VT220: return format_to(ctx.out(), "VT220");
+                case terminal::VTType::VT240: return format_to(ctx.out(), "VT240");
+                case terminal::VTType::VT320: return format_to(ctx.out(), "VT320");
+                case terminal::VTType::VT330: return format_to(ctx.out(), "VT330");
+                case terminal::VTType::VT340: return format_to(ctx.out(), "VT340");
+                case terminal::VTType::VT420: return format_to(ctx.out(), "VT420");
+                case terminal::VTType::VT510: return format_to(ctx.out(), "VT510");
+                case terminal::VTType::VT520: return format_to(ctx.out(), "VT520");
+                case terminal::VTType::VT525: return format_to(ctx.out(), "VT525");
+            }
+            return format_to(ctx.out(), "INVALID-{}", static_cast<unsigned>(_id));
+        }
+    };
+} // }}}