+++ Changelog.md
@@ -23,6 +23,7 @@
 - Adds experimental text reflow.
 - Adds OpenFileManager action to configuration.
 - Adds terminal identification environment variables `TERMINAL_NAME`, `TERMINAL_VERSION_TRIPLE` and `TERMINAL_VERSION_STRING`.
+- Adds config option `profile.*.terminal_id: STR` to set the terminal identification to one of VT100, VT220, VT340, etc.
 - Adds config option `profile.*.maximized: BOOL` to indicate maximized state during profile activation.
 - Adds config option `profile.*.fullscreen: BOOL` to indicate fullscreen state during profile activation.
 - Adds config option `profile.*.font.only_monospace: BOOL` to indicate that only monospace fonts may be used.